Run `mypy_clean_slate`

+++ glass/shells.py
@@ -54,7 +54,7 @@ import numpy.typing as npt
 from glass.core.array import ndinterp
 
 if TYPE_CHECKING:
-    from cosmology import Cosmology
+    from cosmology import Cosmology  # type: ignore[import-untyped]
 
 # types
 ArrayLike1D = Union[Sequence[float], npt.ArrayLike]
@@ -63,17 +63,17 @@ WeightFunc = Callable[[ArrayLike1D], npt.ArrayLike]
 
 def distance_weight(z: npt.ArrayLike, cosmo: Cosmology) -> npt.ArrayLike:
     """Uniform weight in comoving distance."""
-    return 1 / cosmo.ef(z)
+    return 1 / cosmo.ef(z)  # type: ignore[no-any-return]
 
 
 def volume_weight(z: npt.ArrayLike, cosmo: Cosmology) -> npt.ArrayLike:
     """Uniform weight in comoving volume."""
-    return cosmo.xm(z) ** 2 / cosmo.ef(z)
+    return cosmo.xm(z) ** 2 / cosmo.ef(z)  # type: ignore[no-any-return]
 
 
 def density_weight(z: npt.ArrayLike, cosmo: Cosmology) -> npt.ArrayLike:
     """Uniform weight in matter density."""
-    return cosmo.rho_m_z(z) * cosmo.xm(z) ** 2 / cosmo.ef(z)
+    return cosmo.rho_m_z(z) * cosmo.xm(z) ** 2 / cosmo.ef(z)  # type: ignore[no-any-return]
 
 
 class RadialWindow(NamedTuple):
@@ -164,23 +164,23 @@ def tophat_windows(
     :ref:`user-window-functions`
 
     """
-    if len(zbins) < 2:
+    if len(zbins) < 2:  # type: ignore[arg-type]
         msg = "zbins must have at least two entries"
         raise ValueError(msg)
-    if zbins[0] != 0:
+    if zbins[0] != 0:  # type: ignore[index]
         warnings.warn(
             "first tophat window does not start at redshift zero", stacklevel=2
         )
 
     wht: WeightFunc
-    wht = weight if weight is not None else np.ones_like
+    wht = weight if weight is not None else np.ones_like  # type: ignore[assignment]
     ws = []
-    for zmin, zmax in zip(zbins, zbins[1:]):
+    for zmin, zmax in zip(zbins, zbins[1:]):  # type: ignore[arg-type, index]
         n = max(round((zmax - zmin) / dz), 2)
-        z = np.linspace(zmin, zmax, n)
+        z = np.linspace(zmin, zmax, n)  # type: ignore[arg-type]
         w = wht(z)
-        zeff = np.trapz(w * z, z) / np.trapz(w, z)  # type: ignore[attr-defined]
-        ws.append(RadialWindow(z, w, zeff))
+        zeff = np.trapz(w * z, z) / np.trapz(w, z)  # type: ignore[operator] # type: ignore[attr-defined]
+        ws.append(RadialWindow(z, w, zeff))  # type: ignore[arg-type]
     return ws
 
 
@@ -223,25 +223,25 @@ def linear_windows(
     :ref:`user-window-functions`
 
     """
-    if len(zgrid) < 3:
+    if len(zgrid) < 3:  # type: ignore[arg-type]
         msg = "nodes must have at least 3 entries"
         raise ValueError(msg)
-    if zgrid[0] != 0:
+    if zgrid[0] != 0:  # type: ignore[index]
         warnings.warn("first triangular window does not start at z=0", stacklevel=2)
 
     ws = []
-    for zmin, zmid, zmax in zip(zgrid, zgrid[1:], zgrid[2:]):
+    for zmin, zmid, zmax in zip(zgrid, zgrid[1:], zgrid[2:]):  # type: ignore[arg-type, index]
         n = max(round((zmid - zmin) / dz), 2) - 1
         m = max(round((zmax - zmid) / dz), 2)
         z = np.concatenate(
-            [np.linspace(zmin, zmid, n, endpoint=False), np.linspace(zmid, zmax, m)],
+            [np.linspace(zmin, zmid, n, endpoint=False), np.linspace(zmid, zmax, m)],  # type: ignore[arg-type]
         )
         w = np.concatenate(
             [np.linspace(0.0, 1.0, n, endpoint=False), np.linspace(1.0, 0.0, m)],
         )
         if weight is not None:
             w *= weight(z)
-        ws.append(RadialWindow(z, w, zmid))
+        ws.append(RadialWindow(z, w, zmid))  # type: ignore[arg-type]
     return ws
 
 
@@ -285,25 +285,25 @@ def cubic_windows(
     :ref:`user-window-functions`
 
     """
-    if len(zgrid) < 3:
+    if len(zgrid) < 3:  # type: ignore[arg-type]
         msg = "nodes must have at least 3 entries"
         raise ValueError(msg)
-    if zgrid[0] != 0:
+    if zgrid[0] != 0:  # type: ignore[index]
         warnings.warn("first cubic spline window does not start at z=0", stacklevel=2)
 
     ws = []
-    for zmin, zmid, zmax in zip(zgrid, zgrid[1:], zgrid[2:]):
+    for zmin, zmid, zmax in zip(zgrid, zgrid[1:], zgrid[2:]):  # type: ignore[arg-type, index]
         n = max(round((zmid - zmin) / dz), 2) - 1
         m = max(round((zmax - zmid) / dz), 2)
         z = np.concatenate(
-            [np.linspace(zmin, zmid, n, endpoint=False), np.linspace(zmid, zmax, m)],
+            [np.linspace(zmin, zmid, n, endpoint=False), np.linspace(zmid, zmax, m)],  # type: ignore[arg-type]
         )
         u = np.linspace(0.0, 1.0, n, endpoint=False)
         v = np.linspace(1.0, 0.0, m)
         w = np.concatenate([u**2 * (3 - 2 * u), v**2 * (3 - 2 * v)])
         if weight is not None:
             w *= weight(z)
-        ws.append(RadialWindow(z, w, zmid))
+        ws.append(RadialWindow(z, w, zmid))  # type: ignore[arg-type]
     return ws
 
 
@@ -344,7 +344,7 @@ def restrict(
     """
     z_ = np.compress(np.greater(z, w.za[0]) & np.less(z, w.za[-1]), z)
     zr = np.union1d(w.za, z_)
-    fr = ndinterp(zr, z, f, left=0.0, right=0.0) * ndinterp(zr, w.za, w.wa)
+    fr = ndinterp(zr, z, f, left=0.0, right=0.0) * ndinterp(zr, w.za, w.wa)  # type: ignore[operator]
     return zr, fr
 
 
@@ -455,7 +455,7 @@ def partition(
     except KeyError:
         msg = f"invalid method: {method}"
         raise ValueError(msg) from None
-    return partition_method(z, fz, shells)
+    return partition_method(z, fz, shells)  # type: ignore[no-any-return]
 
 
 def partition_lstsq(
@@ -481,7 +481,7 @@ def partition_lstsq(
     dz = np.gradient(zp)
 
     # create the window function matrix
-    a = [np.interp(zp, za, wa, left=0.0, right=0.0) for za, wa, _ in shells]
+    a = [np.interp(zp, za, wa, left=0.0, right=0.0) for za, wa, _ in shells]  # type: ignore[arg-type]
     a /= np.trapz(a, zp, axis=-1)[..., None]  # type: ignore[attr-defined]
     a = a * dz
 
@@ -491,14 +491,14 @@ def partition_lstsq(
 
     # append a constraint for the integral
     mult = 1 / sumtol
-    a = np.concatenate([a, mult * np.ones((len(shells), 1))], axis=-1)
-    b = np.concatenate([b, mult * np.reshape(np.trapz(fz, z), (*dims, 1))], axis=-1)
+    a = np.concatenate([a, mult * np.ones((len(shells), 1))], axis=-1)  # type: ignore[assignment]
+    b = np.concatenate([b, mult * np.reshape(np.trapz(fz, z), (*dims, 1))], axis=-1)  # type: ignore[attr-defined]
 
     # now a is a matrix of shape (len(shells), len(zp) + 1)
     # and b is a matrix of shape (*dims, len(zp) + 1)
     # need to find weights x such that b == x @ a over all axes of b
     # do the least-squares fit over partially flattened b, then reshape
-    x = np.linalg.lstsq(a.T, b.reshape(-1, zp.size + 1).T, rcond=None)[0]
+    x = np.linalg.lstsq(a.T, b.reshape(-1, zp.size + 1).T, rcond=None)[0]  # type: ignore[attr-defined, union-attr]
     x = x.T.reshape(*dims, len(shells))
     # roll the last axis of size len(shells) to the front
     return np.moveaxis(x, -1, 0)
@@ -535,7 +535,7 @@ def partition_nnls(
     dz = np.gradient(zp)
 
     # create the window function matrix
-    a = [np.interp(zp, za, wa, left=0.0, right=0.0) for za, wa, _ in shells]
+    a = [np.interp(zp, za, wa, left=0.0, right=0.0) for za, wa, _ in shells]  # type: ignore[arg-type]
     a /= np.trapz(a, zp, axis=-1)[..., None]  # type: ignore[attr-defined]
     a = a * dz
 
@@ -545,15 +545,15 @@ def partition_nnls(
 
     # append a constraint for the integral
     mult = 1 / sumtol
-    a = np.concatenate([a, mult * np.ones((len(shells), 1))], axis=-1)
-    b = np.concatenate([b, mult * np.reshape(np.trapz(fz, z), (*dims, 1))], axis=-1)
+    a = np.concatenate([a, mult * np.ones((len(shells), 1))], axis=-1)  # type: ignore[assignment]
+    b = np.concatenate([b, mult * np.reshape(np.trapz(fz, z), (*dims, 1))], axis=-1)  # type: ignore[attr-defined]
 
     # now a is a matrix of shape (len(shells), len(zp) + 1)
     # and b is a matrix of shape (*dims, len(zp) + 1)
     # for each dim, find non-negative weights x such that b == a.T @ x
 
     # reduce the dimensionality of the problem using a thin QR decomposition
-    q, r = np.linalg.qr(a.T)
+    q, r = np.linalg.qr(a.T)  # type: ignore[attr-defined]
     y = np.einsum("ji,...j", q, b)
 
     # for each dim, find non-negative weights x such that y == r @ x
@@ -580,7 +580,7 @@ def partition_restrict(
 
 def redshift_grid(
     zmin: float, zmax: float, *, dz: float | None = None, num: int | None = None
-) -> npt.NDArray[float]:
+) -> npt.NDArray[np.float64]:
     """Redshift grid with uniform spacing in redshift."""
     if dz is not None and num is None:
         z = np.arange(zmin, np.nextafter(zmax + dz, zmax), dz)
@@ -599,7 +599,7 @@ def distance_grid(
     *,
     dx: float | None = None,
     num: int | None = None,
-) -> npt.NDArray[float]:
+) -> npt.NDArray[np.float64]:
     """Redshift grid with uniform spacing in comoving distance."""
     xmin, xmax = cosmo.dc(zmin), cosmo.dc(zmax)
     if dx is not None and num is None:
@@ -609,7 +609,7 @@ def distance_grid(
     else:
         msg = 'exactly one of "dx" or "num" must be given'
         raise ValueError(msg)
-    return cosmo.dc_inv(x)
+    return cosmo.dc_inv(x)  # type: ignore[no-any-return]
 
 
 def combine(
@@ -651,11 +651,11 @@ def combine(
     return sum(
         np.expand_dims(weight, -1)
         * np.interp(
-            z,
+            z,  # type: ignore[arg-type]
             shell.za,
             shell.wa / np.trapz(shell.wa, shell.za),  # type: ignore[attr-defined]
             left=0.0,
             right=0.0,
         )
-        for shell, weight in zip(shells, weights)
+        for shell, weight in zip(shells, weights)  # type: ignore[arg-type]
     )
